refactor(site-header-harmonised): specs file, test and story adapted - TWIG-151
