+++ src/ec/packages/ec-component-site-header-harmonised/demo/data--group1.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-site-header-harmonised/demo/data--group1';
+import adapter from '../adapter';
+
+export default adapter(specs);
