+++ src/ec/packages/ec-component-site-header-harmonised/demo/data--group2.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-site-header-harmonised/demo/data--group2';
+import adapter from '../adapter';
+
+export default adapter(specs);
