+++ src/ec/packages/ec-component-site-header-harmonised/site-header-harmonised.story.js
@@ -1,3 +1,4 @@
+import merge from 'deepmerge';
 import { storiesOf } from '@storybook/html';
 import { withNotes } from '@ecl-twig/storybook-addon-notes';
 import withCode from '@ecl-twig/storybook-addon-code';
@@ -5,7 +6,8 @@ import withCode from '@ecl-twig/storybook-addon-code';
 import defaultSprite from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 import logo from '@ecl/ec-resources-logo/logo--en.svg';
 import siteHeaderHarmonised from './ecl-site-header-harmonised.html.twig';
-import { dataGroup1, dataGroup2 } from './demo/data';
+import dataGroup1 from './demo/data--group1';
+import dataGroup2 from './demo/data--group2';
 import notes from './README.md';
 
 storiesOf('Components/Site Headers/Harmonised', module)
@@ -13,25 +15,33 @@ storiesOf('Components/Site Headers/Harmonised', module)
   .addDecorator(withCode)
   .add(
     'group 1',
-    () => {
-      dataGroup1.group = 'group1';
-      dataGroup1.logged = true;
-      dataGroup1.icon_file_path = defaultSprite;
-      dataGroup1.logo.src = logo;
-      return siteHeaderHarmonised(dataGroup1);
-    },
+    () =>
+      siteHeaderHarmonised(
+        merge(dataGroup1, {
+          logo: {
+            src: logo,
+          },
+          group: 'group1',
+          logged: true,
+          icon_file_path: defaultSprite,
+        })
+      ),
     {
       notes: { markdown: notes, json: dataGroup1 },
     }
   )
   .add(
     'group 2',
-    () => {
-      dataGroup2.group = 'group2';
-      dataGroup2.icon_file_path = defaultSprite;
-      dataGroup2.logo.src = logo;
-      return siteHeaderHarmonised(dataGroup2);
-    },
+    () =>
+      siteHeaderHarmonised(
+        merge(dataGroup2, {
+          logo: {
+            src: logo,
+          },
+          group: 'group2',
+          icon_file_path: defaultSprite,
+        })
+      ),
     {
       notes: { markdown: notes, json: dataGroup2 },
     }
